Fixed ILImporter calling ImportStoreElement with Int32 instead of Int64 for i8.

+++ src/Common/src/TypeSystem/IL/ILImporter.cs
@@ -718,7 +718,7 @@ namespace Internal.IL
                         ImportStoreElement(WellKnownType.Int32);
                         break;
                     case ILOpcode.stelem_i8:
-                        ImportStoreElement(WellKnownType.Int32);
+                        ImportStoreElement(WellKnownType.Int64);
                         break;
                     case ILOpcode.stelem_r4:
                         ImportStoreElement(WellKnownType.Single);
